# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BiblioCN Public API (web to agent)
  version: 0.2.0
  description: '三服务再架构的公共 API 契约 (前端 to Python agent 后端)。 单一真源 (Codex-10): 前后端类型由本文件生成, CI 用响应快照防业务语义漂移。
    artifact 带 schemaVersion (Codex-2/Codex-10)。corpus 走显式状态机, 用 oneOf+discriminator 按状态强建模 (Codex 二审
    step1-P1)。错误响应收窄到固定 code, 防契约漂移。

    '
servers:
- url: http://localhost:8000
  description: 本地 agent 服务
paths:
...
  /projects/{pid}/corpus/{cid}/gaps:discover:
    post:
      operationId: discoverGaps
      summary: 启动 GAP 发现 run（scratchpad 编排，异步）
      description: '以 PaperSummary（map 阶段产物）为输入，gap-finder subagent 用 scratchpad 迭代产出 GapCandidate（concept/method/theory
        三 lens）。异步：返回 run_id， 前端轮询 GET .../scratchpad 看实时累积。HITL：不自动定稿。

        '
      tags:
      - research
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
      - name: cid
        in: path
        required: true
        description: R 分析语料 id（rCorpusId 字符串）
        schema:
          type: string
          minLength: 1
          pattern: ^[A-Za-z0-9._-]+$
      responses:
        '202':
          description: 已受理，GAP 发现 run 后台运行中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GapDiscoverAccepted'
        '404':
          description: 项目或语料不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrNotFound'
  /projects/{pid}/agent/runs/{rid}/scratchpad:
    get:
      operationId: getScratchpad
      summary: 拉取本 run 实时 scratchpad（HITL 视图/轮询）
      description: '类 harness 工作记忆：agent 在一次 GAP run 内反复读写的结构化条目（GapCandidate）。 前端像「研究笔记本」实时渲染，并喂回 LLM
        下一轮、作为价值核验输入。

        '
      tags:
      - research
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
      - name: rid
        in: path
        required: true
        description: GAP 发现 run id
        schema:
          type: string
          minLength: 1
          pattern: ^[A-Za-z0-9._-]+$
      responses:
        '200':
          description: 本 run 当前 scratchpad 快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScratchpadState'
        '404':
          description: run 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrNotFound'
  /projects/{pid}/gaps/{gap_id}:verify:
    post:
      operationId: verifyGap
      summary: 启动该 GAP 的价值核验（反向检索证伪 + 计量结构佐证，异步）
      description: 'value-evidence subagent 攒 EvidencePack（反向检索命中 + 计量结构指标）， 确定性 resolver 出 ValueVerdict（decided_by=deterministic）。异步：返回
        verify_run_id。

        '
      tags:
      - research
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
      - name: gap_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          pattern: ^[A-Za-z0-9._-]+$
      responses:
        '202':
          description: 已受理，价值核验 run 后台运行中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GapVerifyAccepted'
        '404':
          description: GAP 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrNotFound'
  /projects/{pid}/gaps/{gap_id}/verdict:
    get:
      operationId: getGapVerdict
      summary: 取价值裁决 + 证据包
      description: 'ValueVerdict 由确定性 resolver 出（decided_by 恒为 deterministic）； 附带攒证的 EvidencePack 供透明审阅（反向检索命中对比、阈值、计量结构）。

        '
      tags:
      - research
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
      - name: gap_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          pattern: ^[A-Za-z0-9._-]+$
      responses:
        '200':
          description: 价值裁决（含证据包）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GapVerdictResult'
        '404':
          description: 裁决尚未产生或 GAP 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrNotFound'
  /projects/{pid}/gaps/{gap_id}:
    patch:
      operationId: patchGap
      summary: HITL 人工 accept/reject/revise
      description: '人在环上裁决：accept/reject 改 status；revise 回写 statement（并置回 draft 以便重核）。 决策留痕进 runlog。绝不自动定稿（HITL
        红线）。

        '
      tags:
      - research
      parameters:
      - name: pid
        in: path
        required: true
        schema:
          type: integer
      - name: gap_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          pattern: ^[A-Za-z0-9._-]+$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GapPatchRequest'
      responses:
        '200':
          description: 更新后的 GapCandidate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GapCandidate'
        '400':
          description: action/statement 校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrValidation'
        '404':
          description: GAP 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrNotFound'
...
components:
...
  schemas:
...
    ApiErrorBase:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          enum:
          - VALIDATION_ERROR
          - CORPUS_NOT_FOUND
          - CORPUS_NOT_READY
          - PARSE_FAILED
          - ANALYSIS_FAILED
          - R_SERVICE_UNAVAILABLE
          - PAYLOAD_TOO_LARGE
          - UNSUPPORTED_FILE
          - INTERNAL
        message:
          type: string
          description: 人类可读, 已脱敏 (Codex-18)
        details:
          type: object
          additionalProperties: true
          nullable: true
...
    ErrValidation:
      allOf:
      - $ref: '#/components/schemas/ApiErrorBase'
      - type: object
        properties:
          code:
            const: VALIDATION_ERROR
    ErrNotFound:
      allOf:
      - $ref: '#/components/schemas/ApiErrorBase'
      - type: object
        properties:
          code:
            const: CORPUS_NOT_FOUND
...
    GapLens:
      type: string
      description: GAP 透镜（概念/方法/理论；领域无关）
      enum:
      - concept
      - method
      - theory
    GapStatus:
      type: string
      description: GAP 生命周期（HITL 流转 draft→verified→accepted/rejected）
      enum:
      - draft
      - verified
      - accepted
      - rejected
    GapSupportingPaper:
      type: object
      description: 支撑该 GAP 的论文证据（带源坐标 anchor_id，可经 EvidenceResolver 回定位原文块）
      required:
      - paper_id
      - anchor_id
      - quote
      properties:
        paper_id:
          type: integer
          description: 语料论文 DB 主键
        anchor_id:
          type: string
          description: 源坐标锚点 id
        quote:
          type: string
          description: 逐字保留的支撑引文（绝不顺手纠正）
    GapCounterEvidence:
      type: object
      description: 反证/张力证据（与 GAP 论断相左的发现）
      required:
      - paper_id
      - anchor_id
      - note
      properties:
        paper_id:
          type: integer
        anchor_id:
          type: string
        note:
          type: string
          description: 反证说明
    ValueThresholds:
      type: object
      description: 透明阈值（反「LLM 拍脑袋」：裁决依据可见、可配、可按领域传参）
      required:
      - reverse_hit_high
      - reverse_hit_low
      properties:
        reverse_hit_high:
          type: integer
          description: 默认 25；hit_count ≥ 此值判 likely_filled（伪空白）
        reverse_hit_low:
          type: integer
          description: 默认 3；hit_count ≤ 此值反向支持真空白
    ValueVerdict:
      type: object
      description: '研究方向价值裁决。由确定性 resolver 出，decided_by 恒为 deterministic； LLM 只产 statement + 攒 EvidencePack，绝不出
        verdict。

        '
      required:
      - gap_id
      - verdict
      - score
      - thresholds
      - rationale
      - decided_by
      properties:
        gap_id:
          type: string
        verdict:
          type: string
          enum:
          - valuable
          - likely_filled
          - inconclusive
        score:
          type: number
          nullable: true
          description: 量化分（可空：inconclusive 时可为 null）
        thresholds:
          $ref: '#/components/schemas/ValueThresholds'
        rationale:
          type: string
          description: 确定性规则的可读解释
        decided_by:
          type: string
          enum:
          - deterministic
          description: 恒为 deterministic（铁律：裁决由确定性代码出，非 LLM）
    GapCandidate:
      type: object
      description: '结构化 GAP 候选（scratchpad 条目）。confidence 为 LLM 自评仅供排序、非裁决依据； value_verdict 由确定性 resolver
        出，未核验时为 null。

        '
      required:
      - gap_id
      - theme
      - statement
      - lens
      - supporting_papers
      - counter_evidence
      - confidence
      - status
      - value_verdict
      properties:
        gap_id:
          type: string
        theme:
          type: string
          description: 所属主题簇
        statement:
          type: string
          description: GAP 论断（如「X 与 Y 的关系在 Z 情境下未被研究」）
        lens:
          $ref: '#/components/schemas/GapLens'
        supporting_papers:
          type: array
          items:
            $ref: '#/components/schemas/GapSupportingPaper'
        counter_evidence:
          type: array
          items:
            $ref: '#/components/schemas/GapCounterEvidence'
        confidence:
          type: number
          description: LLM 自评（0-1，仅供排序，非裁决依据）
        status:
          $ref: '#/components/schemas/GapStatus'
        value_verdict:
          description: 价值裁决；未核验时为 null
          oneOf:
          - $ref: '#/components/schemas/ValueVerdict'
          - type: 'null'
    ReverseSearchHit:
      type: object
      description: 反向检索命中条目（证伪：已有研究是否已填补该空白）
      required:
      - title
      - year
      - doi
      - relevance
      properties:
        title:
          type: string
        year:
          type: integer
          nullable: true
          description: 出版年（可空）
        doi:
          type: string
          nullable: true
          description: DOI（可空）
        relevance:
          type: number
          description: 与 GAP 论断的相关度（0-1）
    ReverseSearch:
      type: object
      description: 反向检索证据（statement→query→Sciverse/OpenAlex 命中计数，去重后）
      required:
      - query
      - provider
      - hit_count
      - top_hits
      properties:
        query:
          type: string
        provider:
          type: string
          enum:
          - sciverse
          - openalex
        hit_count:
          type: integer
          description: 近 N 年强相关计数（去重后，防同一文献多源重复计数虚高）
        top_hits:
          type: array
          items:
            $ref: '#/components/schemas/ReverseSearchHit'
    BiblioStructure:
      type: object
      description: 计量结构佐证（取自现有 R bibliometric 网络数据，不重算）
      required:
      - metric
      - value
      - interpretation
      - source_view
      properties:
        metric:
          type: string
          enum:
          - cooccurrence_gap
          - low_coupling
        value:
          type: number
        interpretation:
          type: string
          description: 确定性可读解释（如「两核心概念间存在共现断层」）
        source_view:
          type: string
          nullable: true
          description: 取自哪个分析视图（如 conceptual），可空
    EvidencePackSkipped:
      type: object
      description: fail-loud：跳过项显式表达（绝不静默成功空结果）
      required:
      - reason
      properties:
        reason:
          type: string
    EvidencePack:
      type: object
      description: '价值核验 subagent 攒的证据包（工具只规整结构、不算术、不裁决）。 gathered_by 恒为 subagent。

        '
      required:
      - gap_id
      - reverse_search
      - biblio_structure
      - gathered_by
      - skipped
      properties:
        gap_id:
          type: string
        reverse_search:
          $ref: '#/components/schemas/ReverseSearch'
        biblio_structure:
          $ref: '#/components/schemas/BiblioStructure'
        gathered_by:
          type: string
          enum:
          - subagent
        skipped:
          type: array
          items:
            $ref: '#/components/schemas/EvidencePackSkipped'
    GapVerdictResult:
      type: object
      description: 'GET .../verdict 返回体：价值裁决 + 攒证的证据包（§2.1「裁决 + 证据包」）。

        '
      required:
      - gap_id
      - verdict
      - evidence
      properties:
        gap_id:
          type: string
        verdict:
          $ref: '#/components/schemas/ValueVerdict'
        evidence:
          $ref: '#/components/schemas/EvidencePack'
    ScratchpadState:
      type: object
      description: 本 run 实时工作记忆快照（类 harness 工作笔记；HITL 轮询视图）
      required:
      - run_id
      - run_status
      - entries
      - updated_at
      properties:
        run_id:
          type: string
        run_status:
          type: string
          enum:
          - running
          - done
          - failed
          description: '本 run 生命周期（轮询停止信号，codex B1-P1）。running 时前端持续轮询； done/failed 时前端停轮询。fail-loud：失败显式
            failed，绝不静默成 done。

            '
        entries:
          type: array
          items:
            $ref: '#/components/schemas/GapCandidate'
        updated_at:
          type: string
          description: ISO8601 更新时间
    GapDiscoverAccepted:
      type: object
      description: :discover 202 受理体
      required:
      - run_id
      properties:
        run_id:
          type: string
    GapVerifyAccepted:
      type: object
      description: :verify 202 受理体
      required:
      - verify_run_id
      properties:
        verify_run_id:
          type: string
    GapAcceptReject:
      type: object
      description: HITL accept/reject（改 status，不带 statement）
      required:
      - action
      properties:
        action:
          type: string
          enum:
          - accept
          - reject
    GapRevise:
      type: object
      description: HITL revise（必带非空 statement，回写论断并置回 draft 以便重核）
      required:
      - action
      - statement
      properties:
        action:
          type: string
          enum:
          - revise
        statement:
          type: string
          minLength: 1
          description: 改写后的新论断（revise 必填非空）
    GapPatchRequest:
      description: 'HITL 决策请求（codex B1-P2：用 oneOf 在契约层锁死「revise 必带 statement」）。 accept/reject 不带 statement；revise
        必带非空 statement。

        '
      oneOf:
      - $ref: '#/components/schemas/GapAcceptReject'
      - $ref: '#/components/schemas/GapRevise'
...